Start working on array serialization
