+++ src/oatpp-postgresql/mapping/Deserializer.cpp
@@ -25,6 +25,7 @@
 #include "Deserializer.hpp"
 
 #include "Oid.hpp"
+#include "PgArray.hpp"
 #include "oatpp-postgresql/Types.hpp"
 
 #if defined(WIN32) || defined(_WIN32)
@@ -279,7 +280,7 @@ const oatpp::Type* Deserializer::guessAnyType(Oid oid) {
 
     case TIMESTAMPOID: return oatpp::UInt64::Class::getType();
 
-    case FLOAT4ARRAYOID: return oatpp::Vector<Float32>::Class::getType();
+    case FLOAT8ARRAYOID: return oatpp::Vector<Float64>::Class::getType();
 
     case UUIDOID: return oatpp::postgresql::Uuid::Class::getType();
 
@@ -315,18 +316,6 @@ oatpp::Void Deserializer::deserializeUuid(const Deserializer* _this, const InDat
   return postgresql::Uuid((p_char8)data.data);
 
 }
-
-// after https://stackoverflow.com/questions/4016412/postgresqls-libpq-encoding-for-binary-transport-of-array-data
-struct PgArray {
-    int32_t ndim;   // Number of dimensions
-    int32_t _ign;   // offset for data, removed by libpq
-    Oid oid;        // type of element in the array
-
-    // Start of array (1st dimension)
-    int32_t size;   // Number of elements
-    int32_t index;  // Index of first element
-    int32_t elem;   // Beginning of (size, value) elements
-};
 
 oatpp::Void Deserializer::deserializeArray(const Deserializer* _this, const InData& data, const Type* type) {
